Add benchmark results for nVidia Jetson AGX Xavier
Benchmark results have a `-mainline` postfix because this is running nvidia's experimental UEFI firmware
with mainline 5.11.6 kernel, instead of their recommended BSP based on 4.9 kernel.

CPU Frequrency that jetson was running was equal to 2265.6 MHz according to sysfs (all cores).

Board itself should be particularly interesting to compare as Tegra 194 (t194 for short) that is a CPU
used on that board, have 8x nVidia Caramel Core which are thier custom VLIW cores that runs ARM code.

A bit of information can be found here:
https://en.wikichip.org/wiki/nvidia/microarchitectures/carmel

Or
https://docs.nvidia.com/jetson/l4t/index.html#page/Tegra%20Linux%20Driver%20Package%20Development%20Guide/power_management_jetson_xavier.html

Unfortunately I don't know what is the default thermal profile on a mainline kernel, but I assume it's equivalent to MAXN as that's the only mode
where you can achive max clocks on all 8 cores.